Refactoring loadDUNS script to avoid circular referencing
